Move composers to a standalone file.
