+++ mart/attack/threat_model.py
@@ -11,8 +11,6 @@ from typing import Any
 
 import torch
 
-__all__ = ["BatchComposer"]
-
 
 class Constraint(abc.ABC):
     @abc.abstractclassmethod
@@ -97,72 +95,3 @@ class BatchEnforcer(Enforcer):
     ) -> torch.Tensor | tuple:
         for input_adv_i, input_i, target_i in zip(input_adv, input, target):
             self._check_constraints(input_adv_i, input_i, target_i)
-
-
-class Composer(torch.nn.Module, abc.ABC):
-    @abc.abstractclassmethod
-    def forward(
-        self,
-        input: torch.Tensor | tuple,
-        target: torch.Tensor | dict[str, Any] | tuple,
-        perturbation: torch.Tensor | tuple,
-    ) -> torch.Tensor | tuple:
-        raise NotImplementedError
-
-
-class BatchComposer(Composer):
-    def __init__(self, composer: Composer):
-        super().__init__()
-
-        self.composer = composer
-
-    def forward(
-        self,
-        input: torch.Tensor | tuple,
-        target: torch.Tensor | dict[str, Any] | tuple,
-        perturbation: torch.Tensor | tuple,
-        **kwargs,
-    ) -> torch.Tensor | tuple:
-        output = []
-
-        for input_i, target_i, perturbation_i in zip(input, target, perturbation):
-            output_i = self.composer(input_i, target_i, perturbation_i, **kwargs)
-            output.append(output_i)
-
-        if isinstance(input, torch.Tensor):
-            output = torch.stack(output)
-        else:
-            output = tuple(output)
-
-        return output
-
-
-class Additive(Composer):
-    """We assume an adversary adds perturbation to the input."""
-
-    def forward(
-        self,
-        input: torch.Tensor,
-        target: torch.Tensor | dict[str, Any],
-        perturbation: torch.Tensor,
-    ) -> torch.Tensor:
-        return input + perturbation
-
-
-class Overlay(Composer):
-    """We assume an adversary overlays a patch to the input."""
-
-    def forward(
-        self,
-        input: torch.Tensor,
-        target: torch.Tensor | dict[str, Any],
-        perturbation: torch.Tensor,
-    ) -> torch.Tensor:
-        # True is mutable, False is immutable.
-        mask = target["perturbable_mask"]
-
-        # Convert mask to a Tensor with same torch.dtype and torch.device as input,
-        #   because some data modules (e.g. Armory) gives binary mask.
-        mask = mask.to(input)
-
-        return input * (1 - mask) + perturbation * mask
