+++ mart/configs/attack/composer/additive.yaml
@@ -1,1 +1,1 @@
-_target_: mart.attack.threat_model.Additive
+_target_: mart.attack.composer.Additive
