+++ mart/configs/attack/composer/overlay.yaml
@@ -1,1 +1,1 @@
-_target_: mart.attack.threat_model.Overlay
+_target_: mart.attack.composer.Overlay
